game resources refactoring
