--- sources/js/game/levels/index.js
@@ -1,43 +1,0 @@
-import level_01 from 'game/levels/level-01.json';
-import level_02 from 'game/levels/level-02.json';
-import level_03 from 'game/levels/level-03.json';
-import level_04 from 'game/levels/level-04.json';
-import level_05 from 'game/levels/level-05.json';
-import level_06 from 'game/levels/level-06.json';
-import level_07 from 'game/levels/level-07.json';
-import level_08 from 'game/levels/level-08.json';
-import level_09 from 'game/levels/level-09.json';
-import level_10 from 'game/levels/level-10.json';
-import level_11 from 'game/levels/level-11.json';
-import level_12 from 'game/levels/level-12.json';
-import level_13 from 'game/levels/level-13.json';
-import level_14 from 'game/levels/level-14.json';
-import level_15 from 'game/levels/level-15.json';
-import level_16 from 'game/levels/level-16.json';
-import level_17 from 'game/levels/level-17.json';
-import level_18 from 'game/levels/level-18.json';
-import level_19 from 'game/levels/level-19.json';
-import level_20 from 'game/levels/level-20.json';
-import level_21 from 'game/levels/level-21.json';
-import level_22 from 'game/levels/level-22.json';
-import level_23 from 'game/levels/level-23.json';
-import level_24 from 'game/levels/level-24.json';
-import level_25 from 'game/levels/level-25.json';
-import level_26 from 'game/levels/level-26.json';
-import level_27 from 'game/levels/level-27.json';
-import level_28 from 'game/levels/level-28.json';
-import level_29 from 'game/levels/level-29.json';
-import level_30 from 'game/levels/level-30.json';
-import level_31 from 'game/levels/level-31.json';
-import level_32 from 'game/levels/level-32.json';
-
-export default [
-	level_01, level_02, level_03, level_04,
-	level_05, level_06, level_07, level_08,
-	level_09, level_10, level_11, level_12,
-	level_13, level_14, level_15, level_16,
-	level_17, level_18, level_19, level_20,
-	level_21, level_22, level_23, level_24,
-	level_25, level_26, level_27, level_28,
-	level_29, level_30, level_31, level_32,
-];
