+++ sources/js/game/resources/levels/level-06.json
