+++ sources/js/game/resources/levels/level-24.json
