+++ sources/js/game/resources/levels/level-26.json
