+++ sources/js/game/score-view.js
